+++ cmd/mux.go
@@ -5,9 +5,10 @@ import (
 	"os"
 	"regexp"
 
-	"github.com/ahamlinman/slackbridge/internal/childproc"
-	"github.com/ahamlinman/slackio"
 	"github.com/spf13/cobra"
+	"go.alexhamlin.co/slackio"
+
+	"go.alexhamlin.co/slackbridge/internal/childproc"
 )
 
 var muxCmd = &cobra.Command{